+++ tests/test_cleaner.py
@@ -1,5 +1,6 @@
 import unittest
 from copy import copy
+from pathlib import Path
 
 import pandas as pd 
 import numpy as np
@@ -13,7 +14,8 @@ from gtfstk.utilities import *
 from gtfstk.feed import *
 from gtfstk.cleaner import *
 
-cairns = read_gtfs('data/cairns_gtfs.zip')
+DATA_DIR = Path('data')
+cairns = read_gtfs(DATA_DIR/'cairns_gtfs.zip')
 
 
 class TestCleaner(unittest.TestCase):
@@ -42,6 +44,26 @@ class TestCleaner(unittest.TestCase):
         feed.routes = pd.concat([old_routes, f])
         new_routes = prune_dead_routes(feed)
         assert_frame_equal(new_routes, old_routes)
+
+    def test_aggregate_routes(self):
+        feed1 = copy(cairns)
+        # Equalize all route short names
+        feed1.routes['route_short_name'] = 'bingo'
+        feed2 = aggregate_routes(feed1)
+
+        # feed2 should have only one route ID
+        self.assertEqual(feed2.routes.shape[0], 1)
+        
+        # Feeds should have same trip data frames excluding
+        # route IDs
+        feed1.trips['route_id'] = feed2.trips['route_id']
+        self.assertTrue(ut.almost_equal(feed1.trips, feed2.trips))
+
+        # Feeds should have equal attributes excluding
+        # routes and trips data frames
+        feed2.routes = feed1.routes 
+        feed2.trips = feed1.trips
+        self.assertEqual(feed1, feed2)
 
 
 if __name__ == '__main__':